Adding configuration

+++ lib/vagrant-notify/action/stop_server.rb
@@ -7,12 +7,14 @@ module Vagrant
         end
 
         def call(env)
+          @app.call env
+          
+          return if env[:machine].config.notify.enable == false
+
           if env[:machine].state.id == :running
             env[:machine].communicate.sudo('rm /usr/bin/notify-send; exit 0')
             env[:machine].communicate.sudo('mv /usr/bin/{notify-send.bkp,notify-send}; exit 0')
           end
-
-          @app.call env
 
           pid = env[:notify_data][:pid]
